+++ src/js/index.ts
@@ -1,1 +1,2 @@
 import './pages/common';
+import './pages/stations';
